Major rework of the DataObjects specs suite
This commit starts the refactor the entire DO spec suite in order
to create a shared test suite that properly covers drivers and
allows for driver authors to be able to say they're driver works
pretty well if they pass these specs.

Besides the new specs, there is also a bunch of fixes in the drivers
done so behavior is more consistent. Also a slew of edge case bugs
is fixed that even could cause Ruby to crash.

This is by no means a finished set, but I'm confident it's already
at least as good as the old specs, so they have been gutted. Area's
that need attention are quoting / battering it with weird input,
more encoding tests and specified DateTime / Time behavior.