Add encoding support for Ruby 1.9
This also improves encoding handling for 1.8, by using a generic
set of encoding names which are the same as they are used in 1.9.
http://pragdave.blogs.pragprog.com/pragdave/2008/12/names-of-encodings-built-in-to-ruby-19.html
provides a list of the encoding names that are available for Ruby.

This adds support for all encodings from that list that are also
support by MySQL and PostgreSQL.