Avoid a segmentation fault upon a null tuple set being returned.
cCommand_execute_non_query() assumes that when a returned status from
PostgreSQL is PGRES_TUPLES_OK that there is always at least one tuple
in the set. This is not always the case as we have at least one INSERT
query with the RETURNING option which always returns a null set due to
a valid database trigger.