Facilitate nested transactions, and DRY up SQL command invocation.
Note that we avoid bouncing through DO's adapter-specific namespace
(e.g. DataObjects::Mysql::) for SavePoints because so far there's no reason to -
the SQL is identical for the two (initially) supported adapters (MySQL +
PostreSQL).

If/when this is necessary, then an overriding SavePoint class should be added to
each DO adapter's lib/<adapter>/transaction.rb, and SavePoint.create_for_uri
will need to be adjusted to compensate.