Fix bug when using nested transactions in concurrent scenarios.
Background:

DB connections are managed via DataObjects::Pooling, which regulates
thread-concurrent DB access to prevent accidental connection sharing (which can
result in anything from dropped connections to field count mismatch exceptions).
They are "acquire"d from the pool when DO::Connection.new is called, and they
are "release"d to the pool when DO::Connection#close is called.

Because SavePoints only happen inside the context of an individual Transaction,
they inherit their encapsulating Transaction's DB connection.  This is also
necessary because DO::Pooling doesn't support the notion of reference counting:
if SP were to acquire a connection from the pool, it would be a different
connection altogether.  Thus, the connection inheritance necessarily sidesteps
the acquisition portion of DO::Pooling.

In a 1 TXN + 1 SP in a multi-threaded scenario:

Due to a bug in the Pooling layer, unwinding the two would actually result in
the connection instance being added back into the @available pool *twice*.  This
is because (1) DM::Transaction#close invokes DO::Connection#close - which
effects the "release" portion of DO::Pooling - and SP inherits TXN's #close, and
(2) DO::Pooling#release blindly adds the instance back to the pool (fixed in
previous commit).  All of which effectively enables a race between that Thread
and any others in the process of acquiring/using a connection, making it
possible for threads to step on each other's connection.

The fix is to no-op SP#close, which means the connection will remain unavailable
to other threads until the encapsulating TXN#close is invoked.

Yugh.