Fix hangs in specs
The issue with the hangs was because how before / after blocks work
in shared specs. Apparently they are shared across different shared
blocks so they were run twice which caused additional connection
objects to be allocated which weren't referenced afterwards and
couldn't be cleaned up...

Reordering the before blocks setting up the connections fixed this
issue.