Fix more Connection Pool leaks in tests.
Building on Dirkjan's d19180fe, this solves a few more connection pool leaks by
ensuring that all `before'/`after's that acquire/release from the pool are
inside `describe' blocks (they shouldn't be immediately inside `shared' blocks).

Currently triggering via fails on a Linux build box during the "parallel query"
spec, which launches a 4 threads that use SQL to sleep 1, and expects them all
to finish within 2 seconds.

The few of us testing haven't seen this yet probably due to the subtle
Dir[**/spec_*.rb] path-glob-order-relies-on-inode-order problem: depending on
the order the spec files are run, the "parallel query" spec will either come
earlier or later, and the earlier it is, the less likely it is to trigger (being
the largest one-time consumer of the 8 maximum allowed pooled connections).