Give failures to connect one last shot with UTF-8 encoding.
Previously we were more selective about when we would attempt this; bad
encodings were indicated through text inside the exception text.  However this
seems to have changed lately, and failures to connect are surfaced without much
more detail about the underlying why.

So we just try once more with UTF-8 encoding (default) no matter what.  This is
what the do_{adapter} specs suggest is the intended behaviour anyway.  If the
connection failed for a reason other than encoding, then it'll fail again and
the user will end up with the same error anyway.