Don't use specialized code for creating Date / DateTime's
This code is pretty fragile and breaks if you for example
use something like home_run in certain edge cases. This
is a performance optimization in the first place and if
people want / need this optimization they can either use
something like home_run to replace Date with a faster version
or upgrade to 1.9.3 which has a faster Date anyway.